+++ Common/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [3.0.9] - 2024-06-01
+- Fixed support for Microsoft.CodeAnalysis 4.11.0 (preview versions)
+- Fixed support for Microsoft.CodeAnalysis 4.10.0 (preview versions)
+
 ## [3.0.8] - 2023-10-14
 - Fixed support for Microsoft.CodeAnalysis 4.9.0 (preview versions)